organize spatial hashing
